+++ modules/@apostrophecms/doc-type/ui/apos/components/AposDocEditor.vue
@@ -350,8 +350,8 @@ export default {
       type: this.$t(this.moduleOptions.label)
     };
     if (this.docId) {
-      this.evaluateConditions();
       await this.loadDoc();
+      this.evaluateConditions();
       try {
         if (this.manuallyPublished) {
           this.published = await apos.http.get(this.getOnePath, {