+++ modules/@apostrophecms/schema/ui/apos/components/AposInputPassword.vue
@@ -15,6 +15,7 @@
           :required="field.required"
           :id="uid"
           :tabindex="tabindex"
+          :autocomplete="field.autocomplete"
           @keydown.enter="$emit('return')"
         >
       </div>
@@ -26,6 +27,7 @@
 import AposInputPasswordLogic from '../logic/AposInputPassword';
 export default {
   name: 'AposInputPassword',
-  mixins: [ AposInputPasswordLogic ]
+  mixins: [ AposInputPasswordLogic ],
+  emits: [ 'return' ]
 };
 </script>